+++ demisto_sdk/commands/format/update_generic_json.py
@@ -1,9 +1,9 @@
 from distutils.version import LooseVersion
 
+import click
 import ujson
 import yaml
-from demisto_sdk.commands.common.tools import (LOG_COLORS, print_color,
-                                               print_error)
+from demisto_sdk.commands.common.tools import print_error
 from demisto_sdk.commands.format.format_constants import (
     ARGUMENTS_DEFAULT_VALUES, TO_VERSION_5_9_9)
 from demisto_sdk.commands.format.update_generic import BaseUpdate
@@ -17,19 +17,22 @@ class BaseUpdateJSON(BaseUpdate):
             data (Dict): JSON file data arranged in a Dict.
     """
 
-    def __init__(self, input: str = '', output: str = '', path: str = '', from_version: str = '', no_validate: bool = False):
-        super().__init__(input=input, output=output, path=path, from_version=from_version, no_validate=no_validate)
+    def __init__(self, input: str = '', output: str = '', path: str = '', from_version: str = '', no_validate: bool = False, verbose: bool = False):
+        super().__init__(input=input, output=output, path=path, from_version=from_version, no_validate=no_validate,
+                         verbose=verbose)
 
     def set_default_values_as_needed(self):
         """Sets basic arguments of reputation commands to be default, isArray and required."""
-        print('Updating required default values')
+        if self.verbose:
+            click.echo('Updating required default values')
         for field in ARGUMENTS_DEFAULT_VALUES:
             if self.__class__.__name__ in ARGUMENTS_DEFAULT_VALUES[field][1]:
                 self.data[field] = ARGUMENTS_DEFAULT_VALUES[field][0]
 
     def save_json_to_destination_file(self):
         """Save formatted JSON data to destination file."""
-        print_color(f'Saving output JSON file to {self.output_file}', LOG_COLORS.WHITE)
+        if self.source_file != self.output_file:
+            click.secho(f'Saving output JSON file to {self.output_file}', fg='white')
         with open(self.output_file, 'w') as file:
             ujson.dump(self.data, file, indent=4)
 
@@ -46,13 +49,15 @@ class BaseUpdateJSON(BaseUpdate):
         Relevant for old entities such as layouts and classifiers.
         """
         if not self.data.get('toVersion') or LooseVersion(self.data.get('toVersion', '99.99.99')) >= TO_VERSION_5_9_9:
-            print('Setting toVersion field')
+            if self.verbose:
+                click.echo('Setting toVersion field')
             self.data['toVersion'] = TO_VERSION_5_9_9
 
     def set_description(self):
         """Add an empty description to file root."""
         if 'description' not in self.data:
-            print('Adding empty descriptions to root')
+            if self.verbose:
+                click.echo('Adding empty descriptions to root')
             self.data['description'] = ''
 
     def remove_null_fields(self):
@@ -68,7 +73,8 @@ class BaseUpdateJSON(BaseUpdate):
     def update_id(self, field='name'):
         """Updates the id to be the same as the provided field ."""
 
-        print('Updating ID')
+        if self.verbose:
+            click.echo('Updating ID')
         if field not in self.data:
             print_error(f'Missing {field} field in file {self.source_file} - add this field manually')
             return