+++ CHANGELOG.md
@@ -9,10 +9,13 @@
 * Updated the missing release notes error message (RN106) in the **Validate** command.
 * Fixed an issue where **Validate** would return missing release notes when two packs with the same substring existed in the modified files.
 * Fixed an issue where **update-release-notes** would add duplicate release notes when two packs with the same substring existed in the modified files.
+* Fixed an issue where **update-release-notes** would fail to bump new versions if the feature branch was out of sync with the master branch.
 * Fixed an issue where a non-descriptive error would be returned when giving the **update-release-notes** command a pack which can not be found.
 * Added dependencies check for *widgets* in **find-dependencies** command.
 * Added support for *update-docker* in **format** command.
 * Added a `json-to-outputs` flag to the **run** command.
+* Added a verbose (`-v`) flag to **format** command.
+* Fixed an issue where **download** added the prefix "playbook-" to the name of playbooks.
 
 # 1.1.10
 * Updated the **init** command. Relevant only when passing the *--contribution* argument.